[Gold IV] Title: 최단경로, Time: 896 ms, Memory: 124848 KB -BaekjoonHub

+++ 백준/Gold/1753. 최단경로/최단경로.java
@@ -11,7 +11,7 @@ public class Main {
         StringTokenizer st = new StringTokenizer(br.readLine());
         int V = Integer.parseInt(st.nextToken());
         int E = Integer.parseInt(st.nextToken());
-        Map<Integer, Integer>[] mapArr = new Map[V + 1];
+        Map<Integer, Byte>[] mapArr = new Map[V + 1];
         int[] weight = new int[V + 1];
         for (int i = 0; i < V + 1; i++) {
             mapArr[i] = new HashMap<>();
@@ -24,9 +24,9 @@ public class Main {
             st = new StringTokenizer(br.readLine());
             int u = Integer.parseInt(st.nextToken());   // 시작
             int v = Integer.parseInt(st.nextToken());   // 도착
-            int w = Integer.parseInt(st.nextToken());   // 가중치(10 이하)
+            byte w = Byte.parseByte(st.nextToken());   // 가중치(10 이하)
 
-            int value = mapArr[u].getOrDefault(v, Integer.MAX_VALUE);
+            byte value = mapArr[u].getOrDefault(v, Byte.MAX_VALUE);
             if (w < value) {
                 mapArr[u].put(v, w);
             }
@@ -37,10 +37,10 @@ public class Main {
         while (!queue.isEmpty()) {
             int point = queue.poll();
             int prevVal = weight[point];
-            for (Map.Entry<Integer, Integer> entry : mapArr[point].entrySet()) {
+            for (Map.Entry<Integer, Byte> entry : mapArr[point].entrySet()) {
                 int key = entry.getKey();
                 int value = prevVal + entry.getValue();
-                if(weight[key] > value) {
+                if (weight[key] > value) {
                     weight[key] = value;
                     queue.add(key);
                 }
@@ -48,7 +48,7 @@ public class Main {
         }
 
         StringBuilder sb = new StringBuilder();
-        for(int i=1; i< weight.length; i++) {
+        for (int i = 1; i < weight.length; i++) {
             sb.append(weight[i] == Integer.MAX_VALUE ? "INF" : weight[i]).append('\n');
         }
         System.out.print(sb);
